+++ src/main/java/com/example/springrecipetest/controllers/RecipeController.java
@@ -70,20 +70,7 @@ public class RecipeController {
 
     }
 
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ExceptionHandler(NumberFormatException.class)
-    public ModelAndView handleNumberFormat(Exception exception) {
-        log.error("Handling Number Format exception");
-        log.error(exception.getMessage());
 
-        ModelAndView modelAndView = new ModelAndView();
-
-        modelAndView.setViewName("404error");
-        modelAndView.addObject("exception", exception);
-
-        return modelAndView;
-
-    }
